api/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
attrs==20.3.0
black==20.8b1
click==7.1.2
Django==3.2
django-nose==1.4.7
djangorestframework==3.12.4
iniconfig==1.1.1
mypy-extensions==0.4.3
nose==1.3.7
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest-django==4.2.0
pytz==2021.1
regex==2021.4.4
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
